+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Adds `Overstrike.new`.
 - Adds `Overstrike.noop`.
 - Adds `Overstrike.identity`.
+- Adds `Overstrike.constant`.
 
 ## 0.1.337 (Aug 4, 2014)
 - Adds `Overstrike.assign`.